Make linker generic
